Check for data validation input
if the user adds a check like this in their template, the Campaign Generator will catch this and stop the user from submitting the template/recipient combo.  This will also get captured during a transmission call with similarly bad data.  The user just needs to check for non-existent data fields that are important to the template.  

If the data is there, everything will run correctly.
If the data is missing, output similar to this is generated:
array (
    'message' => 'substitution render error',
    'description' => 'Error while rendering part html: line 231: attempt to call non-existent macro \'crash\'',
    'code' => '3001',
    'part' => 'html',
    'line' => 231,
  ),

This is caught in this program when it checks the first recipient user.  If the first user is good but others are bad, the campaign can be launched and the SparkPost generation system will create 'generation failure events' for any recipients that don't have the needed data.

The data check will also work for any transmission call from other programs.  Good data sets will work, the others will not generate.  This protects the user from getting charged for accidentally sending bad emails with incomplete data sets.  It doesn't protect against everything; like old data, bad data; but it does a decent job of catching stupid stuff :-)