Continued comments for data validation
To protect against selecting the wrong Recipient List for a template and thusly getting charged for bad emails you can protect yourself by adding the following code (or something similar) somewhere in your template.  What this code does is check that the substitution json stanza has the important data needed for this template.  In this example, the template is expecting a list of products, if the 'productlist' stanza is not found, the generator will attempt the 'if crash.....' statement which will cause the generation to fail.

{if !productlist}} <h5 class="hidden">MISSING KEY DATA</h5> {{if crash(onpurpose) }}{{end}}{{end}}

The output from the generation process will be something like this:

array (
    'message' => 'substitution render error',
    'description' => 'Error while rendering part html: line 231: attempt to call non-existent macro \'crash\'',
    'code' => '3001',
    'part' => 'html',
    'line' => 231,
  )

By doing this, the Campaign Generator Application (this application) has one more check to protect against bad emails being sent.  While checking for missing data is optional; you MUST add the following text  "{{if crash(onpurpose) }}{{end}}{{end}}" after your variable check in order for this application to see the problem.  Otherwise, some combinations of template/recipients can be valid in the eye's of the SparkPost generation process, but result in an email with a lot of missing data.